#31 リロードしても従業員一覧が表示されるよう変更
