+++ src/contexts/Auth.tsx
@@ -1,7 +1,6 @@
-import React, { FC, createContext, useState, useEffect, useMemo } from 'react';
+import React, { FC, createContext, useState } from 'react';
 
 import LoginType from '../types/loginType';
-import createAxiosClient from '../api/client';
 
 type ContextType = {
   loggedInType: LoginType | undefined;
@@ -10,60 +9,15 @@ type ContextType = {
   setLoggedInUserId: React.Dispatch<React.SetStateAction<string | undefined>>;
 };
 
-type CheckResponseType = {
-  data: {
-    type: string;
-  };
-};
-
-type UserIdResponseType = {
-  data: {
-    // eslint-disable-next-line
-    user_id: string;
-  };
-};
-
 const AuthContext = createContext({} as ContextType);
 
 const AuthProvider: FC = ({ children }) => {
-  const axiosClient = useMemo(() => createAxiosClient(), []);
-
   const [loggedInType, setLoggedInType] = useState<LoginType | undefined>(
     undefined,
   );
   const [loggedInUserId, setLoggedInUserId] = useState<string | undefined>(
     undefined,
   );
-
-  useEffect(() => {
-    axiosClient('check')
-      .then((res: CheckResponseType) => {
-        if (res.data.type === 'none') {
-          setLoggedInType(undefined);
-        } else {
-          setLoggedInType(res.data.type as LoginType);
-        }
-      })
-      .catch((err) => {
-        // eslint-disable-next-line
-        console.log(err);
-      });
-  }, [axiosClient]);
-
-  useEffect(() => {
-    axiosClient('check/user_id')
-      .then((res: UserIdResponseType) => {
-        if (res.data.user_id === 'none') {
-          setLoggedInUserId(undefined);
-        } else {
-          setLoggedInUserId(res.data.user_id);
-        }
-      })
-      .catch((err) => {
-        // eslint-disable-next-line
-        console.log(err);
-      });
-  }, [axiosClient]);
 
   return (
     <AuthContext.Provider
